correcting the code in accordance with the mentor's comments BJS2-33918

+++ src/test/java/school/faang/BJS2_33918/DataCenterTest.java
@@ -25,7 +25,7 @@ public class DataCenterTest {
 
     @Test
     void subtractionTest() {
-        int result = DataCenter.calculate(NUMBERS, (a, b) -> (int) a - (int) b);
+        int result = DataCenter.calculate(NUMBERS, (a, b) -> a - b);
 
         assertEquals(-13, result);
     }
